collated mock integration tests to use when modifying phase
