+++ autolens/pipeline/phase/phase_imaging.py
@@ -1,21 +1,23 @@
-from astropy import cosmology as cosmo
 import numpy as np
+from astropy import cosmology as cosmo
 
 import autofit as af
-from autolens.array.util import binning_util
 from autolens.array import mask as msk
+from autolens.array.util import binning_util
 from autolens.lens import lens_data as ld
 from autolens.lens import lens_fit
 from autolens.model.galaxy import galaxy as g
 from autolens.pipeline import phase_tagging
-from autolens.pipeline.phase.phase_data import PhaseData
 from autolens.pipeline.phase import phase_extensions
+from autolens.pipeline.phase.phase_data import PhaseData
 from autolens.pipeline.plotters import phase_plotters
+
 
 def default_mask_function(image):
     return msk.Mask.circular(
         shape=image.shape, pixel_scale=image.pixel_scale, sub_size=1, radius_arcsec=3.0
     )
+
 
 def isinstance_or_prior(obj, cls):
     if isinstance(obj, cls):
@@ -24,33 +26,33 @@ def isinstance_or_prior(obj, cls):
         return True
     return False
 
-class PhaseImaging(PhaseData):
 
+class PhaseImaging(PhaseData):
     galaxies = af.PhaseProperty("galaxies")
     hyper_image_sky = af.PhaseProperty("hyper_image_sky")
     hyper_background_noise = af.PhaseProperty("hyper_background_noise")
 
     def __init__(
-        self,
-        phase_name,
-        phase_folders=tuple(),
-        galaxies=None,
-        hyper_image_sky=None,
-        hyper_background_noise=None,
-        optimizer_class=af.MultiNest,
-        cosmology=cosmo.Planck15,
-        sub_size=2,
-        signal_to_noise_limit=None,
-        bin_up_factor=None,
-        psf_shape=None,
-        positions_threshold=None,
-        mask_function=None,
-        inner_mask_radii=None,
-        pixel_scale_interpolation_grid=None,
-        pixel_scale_binned_cluster_grid=None,
-        inversion_uses_border=True,
-        inversion_pixel_limit=None,
-        auto_link_priors=False,
+            self,
+            phase_name,
+            phase_folders=tuple(),
+            galaxies=None,
+            hyper_image_sky=None,
+            hyper_background_noise=None,
+            optimizer_class=af.MultiNest,
+            cosmology=cosmo.Planck15,
+            sub_size=2,
+            signal_to_noise_limit=None,
+            bin_up_factor=None,
+            psf_shape=None,
+            positions_threshold=None,
+            mask_function=None,
+            inner_mask_radii=None,
+            pixel_scale_interpolation_grid=None,
+            pixel_scale_binned_cluster_grid=None,
+            inversion_uses_border=True,
+            inversion_pixel_limit=None,
+            auto_link_priors=False,
     ):
 
         """
@@ -160,7 +162,8 @@ class PhaseImaging(PhaseData):
 
         pixel_scale_binned_grid = self.pixel_scale_binned_grid_from_mask(mask=mask)
 
-        preload_pixelization_grids_of_planes = self.preload_pixelization_grids_of_planes_from_results(results=results)
+        preload_pixelization_grids_of_planes = self.preload_pixelization_grids_of_planes_from_results(
+            results=results)
 
         lens_imaging_data = ld.LensImagingData(
             imaging_data=data,
@@ -222,11 +225,11 @@ class PhaseImaging(PhaseData):
             phase_info.close()
 
     def extend_with_multiple_hyper_phases(
-        self,
-        hyper_galaxy=False,
-        inversion=False,
-        include_background_sky=False,
-        include_background_noise=False,
+            self,
+            hyper_galaxy=False,
+            inversion=False,
+            include_background_sky=False,
+            include_background_noise=False,
     ):
 
         if hyper_galaxy:
@@ -414,7 +417,7 @@ class PhaseImaging(PhaseData):
             """
             if hasattr(self, "hyper_galaxy_image_1d_path_dict"):
                 for galaxy_path, galaxy in instance.path_instance_tuples_for_class(
-                    g.Galaxy
+                        g.Galaxy
                 ):
                     if galaxy_path in self.hyper_galaxy_image_1d_path_dict:
                         galaxy.hyper_model_image_1d = self.hyper_model_image_1d
@@ -422,8 +425,8 @@ class PhaseImaging(PhaseData):
                             galaxy_path
                         ]
                         if (
-                            hasattr(self, "binned_hyper_galaxy_image_1d_path_dict")
-                            and self.binned_hyper_galaxy_image_1d_path_dict is not None
+                                hasattr(self, "binned_hyper_galaxy_image_1d_path_dict")
+                                and self.binned_hyper_galaxy_image_1d_path_dict is not None
                         ):
                             galaxy.binned_hyper_galaxy_image_1d = self.binned_hyper_galaxy_image_1d_path_dict[
                                 galaxy_path
@@ -537,13 +540,13 @@ class PhaseImaging(PhaseData):
 
     class Result(PhaseData.Result):
         def __init__(
-            self,
-            constant,
-            figure_of_merit,
-            previous_variable,
-            gaussian_tuples,
-            analysis,
-            optimizer,
+                self,
+                constant,
+                figure_of_merit,
+                previous_variable,
+                gaussian_tuples,
+                analysis,
+                optimizer,
         ):
             """
             The result of a phase
@@ -646,7 +649,7 @@ class PhaseImaging(PhaseData):
                     minimum_galaxy_value = hyper_minimum_percent * max(galaxy_image_1d)
                     galaxy_image_1d[
                         galaxy_image_1d < minimum_galaxy_value
-                    ] = minimum_galaxy_value
+                        ] = minimum_galaxy_value
 
                 hyper_galaxy_image_1d_path_dict[path] = galaxy_image_1d
 
@@ -712,7 +715,7 @@ class PhaseImaging(PhaseData):
                     )
                     binned_galaxy_image_1d[
                         binned_galaxy_image_1d < minimum_hyper_value
-                    ] = minimum_hyper_value
+                        ] = minimum_hyper_value
 
                     binned_hyper_galaxy_image_path_dict[path] = binned_galaxy_image_1d
