--- test/integration/tests/lens__source/mock_nlo/lens_mass__source_x2__hyper.py
@@ -1,7 +1,0 @@
-from test.integration.tests.lens__source import lens_mass__source_x2__hyper
-from test.integration.tests.runner import run_a_mock
-
-
-class TestCase:
-    def _test_lens_mass__source_x2__hyper(self):
-        run_a_mock(lens_mass__source_x2__hyper)
